backend/requirements.txt
Django==5.1.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
django-widget-tweaks==1.5.0
